requirements.txt
ffmpeg-python~=0.2.0
pillow~=10.4.0
torchvision~=0.19.1